pyproject.toml
[build-system]
requires = [
    "setuptools>=61",
    # 4110 only:
    "cython>=3",
]
build-backend = "setuptools.build_meta"

[project]
version = "1.0.0"
requires-python = ">=3.8"
license = {text = "MIT License"}
name = "in3110_instapy"
description = "Turn a regular image into a gray scale version"
readme = "README.md"
dependencies = [
    "numpy",
    "numba",
    "pillow",
    "line-profiler"
]
[project.optional-dependencies]
test = [
    "pytest",
    "pytest-benchmark"
]

[project.scripts]
instapy = "in3110_instapy.cli:main"
